two reproducible analysis completed
This section of code loads in the csv file from The Washington Post, determines two reproducible questions, and finds the same numbers that the WP found.

The second question: the ratio of arrest in 2014 compared to 2017 was done in three different variables with pipes, though I suspect that it could be done more efficiently but as of right now I do not know how to do that.

Additionally, I used a new function called merge() which seems to be a basic join function. I honestly just typed in merge() and hoped that it would be what I needed but I double checked some documentation and the previously made data frames to make sure I did not lose anything.